client/requirements.txt
cmd2
python-digitalocean
pyyaml
paramiko
setuptools-rust